// src/shared/api/generated/models/updateCompanyDto.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * VIO Stats Service API
 * API для управления компаниями, рекрутерами и вакансиями
 * OpenAPI spec version: 1.0
 */

export interface UpdateCompanyDto {
  /** Название компании */
  name?: string;
  /** Email компании */
  email?: string;
  /** Телефон компании */
  phone?: string;
  /** Адрес компании */
  address?: string;
}